reap: execv(): return error to caller
The error was previously logged and a non-0 exit status returned.

This change revealed a bug in the tests: wait(2) errors out sometimes
with ECHILD. The tests passed because the exit status was non-0.

wait(2) could be returning ECHILD because the subprocess is killed before
doing the exec(). Running reap() attempts to terminate all subprocesses
of the test. Running reap() loops concurrently may race goroutines
fork()'ing to execute the test commmand.